engines/thumbnail_engine.py
"""
THUMBNAIL ENGINE - FLUX.1 Integration
Generates CTR-optimized thumbnails with Hindi text
"""
from pathlib import Path
from loguru import logger
from config import Config


class ThumbnailEngine:
    """Thumbnail generation engine using FLUX.1"""
    
    def __init__(self):
        self.config = Config()
        self.logger = logger
        self.api_key = self.config.FLUX_API_KEY
    
    async def generate(
        self,
        video_path: str,
        topic: str,
        language: str = "hindi"
    ) -> str:
        """
        Generate thumbnail from video
        
        Args:
            video_path: Path to final video
            topic: Video topic
            language: Language for text
            
        Returns:
            Path to thumbnail image
        """
        self.logger.info(f"🖼️  Generating thumbnail...")
        
        # For now, create placeholder
        # TODO: Implement FLUX.1 integration
        self.logger.warning("⚠️  FLUX.1 integration pending - creating placeholder")
        
        return await self._create_placeholder(video_path, topic)
    
    async def _create_placeholder(self, video_path: str, topic: str) -> str:
        """Create placeholder thumbnail from video frame"""
        import subprocess
        
        # Extract middle frame from video
        output_path = self.config.OUTPUT_PATH / f"{Path(video_path).stem}_thumb.jpg"
        
        cmd = [
            self.config.FFMPEG_PATH,
            "-i", video_path,
            "-vf", "select='eq(n,2)'",
            "-vframes", "1",
            "-y",
            str(output_path)
        ]
        
        try:
            subprocess.run(cmd, capture_output=True, text=True)
            self.logger.success(f"✅ Thumbnail created: {output_path.name}")
            return str(output_path)
        except Exception as e:
            self.logger.error(f"Thumbnail extraction failed: {e}")
            # Create dummy image
            from PIL import Image
            img = Image.new('RGB', (1280, 720), color=(50, 50, 100))
            img.save(str(output_path))
            return str(output_path)
